requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
open_clip_torch==2.29.0
torch==2.5.1
torchvision==0.20.1
numpy==1.26.4
pillow==11.0.0
piexif==1.1.3
watchdog==6.0.0
reverse-geocoder==1.5.1
sqlite-vec==0.1.6
python-multipart==0.0.12
